+++ ng-workspace/projects/OpenWaterFoundation/common/ui/dashboard/widget/title/title.module.ts
@@ -1,0 +1,36 @@
+import { NgModule }       from '@angular/core';
+import { CommonModule }   from '@angular/common';
+
+import { ShowdownModule } from 'ngx-showdown';
+
+import { ErrorModule }    from '../error/error.module';
+import { TitleComponent }  from './title.component';
+
+
+@NgModule({
+  declarations: [
+    TitleComponent
+  ],
+  exports: [
+    TitleComponent
+  ],
+  imports: [
+    CommonModule,
+
+    ShowdownModule.forRoot({
+      emoji: true,
+      flavor: 'github',
+      noHeaderId: true,
+      openLinksInNewWindow: true,
+      parseImgDimensions: true,
+      simpleLineBreaks: false,
+      strikethrough: true,
+      tables: true
+    }),
+
+    ErrorModule
+  ],
+  providers: [
+  ]
+})
+export class TitleModule { }